+++ spec/rubocop/cli_spec.rb
@@ -302,9 +302,10 @@ RSpec.describe RuboCop::CLI, :isolated_environment do
         .to eq(['example.rb: Style/LineLength has the wrong ' \
                 'namespace - should be Metrics',
                 ''].join("\n"))
-      # 3 cops were disabled, then 2 were enabled again, so we
-      # should get 2 offenses reported.
+      # 2 real cops were disabled, and 1 that was incorrect
+      # 2 real cops was enabled, but only 1 had been disabled correctly
       expect($stdout.string).to eq(<<-RESULT.strip_indent)
+        #{abs('example.rb')}:6:21: W: Lint/UnneededCopEnableDirective: Unnecessary enabling of Metrics/LineLength.
         #{abs('example.rb')}:7:81: C: Metrics/LineLength: Line is too long. [95/80]
         #{abs('example.rb')}:9:5: C: Style/StringLiterals: Prefer single-quoted strings when you don't need string interpolation or special symbols.
       RESULT